Move admin to admin path
