+++ app/controllers/admin/admin_controller.rb
@@ -1,0 +1,5 @@
+module Admin
+  class AdminController < ApplicationController
+    include AdminAuthorizationFilters
+  end
+end
