--- app/controllers/admin_controller.rb
@@ -1,3 +1,0 @@
-class AdminController < ApplicationController
-  include AdminAuthorizationFilters
-end
