+++ app/views/admin/uploads/complete.html.erb
@@ -2,7 +2,7 @@
   <%= t('page_title.upload') %>
 <% end %>
 
-<%= govuk_back_link(href: "/upload", text: "Back") %>
+<%= govuk_back_link(href: admin_upload_path, text: "Back") %>
 
 <h1 class="govuk-heading-xl">Upload CSV Complete</h1>
