+++ app/views/admin/uploads/new.html.erb
@@ -2,13 +2,13 @@
   <%= t('page_title.upload') %>
 <% end %>
 
-<%= button_to("Log out", "/logout", class: "govuk-back-link", method: :delete) %>
+<%= button_to("Log out", admin_logout_path, class: "govuk-back-link", method: :delete) %>
 
 <h1 class="govuk-heading-xl">Upload CSV</h1>
 
 <div class="govuk-grid-row">
   <div class="govuk-grid-column-one-half">
-    <%= form_for @upload, url: upload_path, multipart: true, builder: GOVUKDesignSystemFormBuilder::FormBuilder do |f| %>
+    <%= form_for @upload, url: admin_upload_path, multipart: true, builder: GOVUKDesignSystemFormBuilder::FormBuilder do |f| %>
       <%= f.govuk_error_summary %>
       <%= f.govuk_file_field :file %>
       <%= f.govuk_submit "Upload" %>
