+++ spec/requests/login_spec.rb
@@ -1,49 +1,40 @@
 RSpec.describe "Login" do
   let!(:user) { create(:user, name: "test_user", password: "mypass") }
 
-  it "does not allow access to the login page if not on the admin subdomain" do
+  it "allows user to login" do
     get admin_path
-    expect(response.body).not_to include("Admin login")
+    expect(response).to be_successful
+    expect(response.body).to include("Admin login")
+
+    post admin_login_path, params: { login: { name: "test_user", password: "mypass" } }
+    expect(response).to redirect_to(admin_upload_path)
   end
 
-  context "when on admin subdomain" do
-    before { use_admin_subdomain }
+  it "loads user into session" do
+    request_login_as("test_user", "mypass")
+    expect(session[:user_id]).to eq user.id
+  end
 
-    it "allows user to login" do
-      get admin_path
-      expect(response).to be_successful
-      expect(response.body).to include("Admin login")
+  it "handles invalid login details" do
+    post admin_login_path, params: { login: { name: "test_user", password: "" } }
+    expect(response.body).to include("There is a problem")
+  end
 
-      post login_path, params: { login: { name: "test_user", password: "mypass" } }
-      expect(response).to redirect_to(upload_path)
-    end
-
-    it "loads user into session" do
+  context "when logged in" do
+    before do
       request_login_as("test_user", "mypass")
-      expect(session[:user_id]).to eq user.id
     end
 
-    it "handles invalid login details" do
-      post login_path, params: { login: { name: "test_user", password: "" } }
-      expect(response.body).to include("There is a problem")
+    it "allows user to logout" do
+      delete admin_logout_path
+      expect(response).to redirect_to(admin_path)
+
+      expect(session[:user_id]).to be_nil
     end
 
-    context "when logged in" do
-      before do
-        request_login_as("test_user", "mypass")
-      end
-
-      it "allows user to logout" do
-        delete logout_path
-        expect(response).to redirect_to(admin_path)
-
-        expect(session[:user_id]).to be_nil
-      end
-
-      it "automatically redirects user away from login page" do
-        get admin_path
-        expect(response).to redirect_to(upload_path)
-      end
+    it "automatically redirects user away from login page" do
+      get admin_path
+      expect(response).to redirect_to(admin_upload_path)
     end
   end
 end
