+++ spec/requests/upload_spec.rb
@@ -1,9 +1,7 @@
 RSpec.describe "Upload" do
-  before { use_admin_subdomain }
-
   context "when not logged in" do
     it "redirects if not logged in" do
-      get upload_path
+      get admin_upload_path
       expect(response).to be_redirect
     end
   end
@@ -15,19 +13,19 @@ RSpec.describe "Upload" do
     end
 
     it "shows upload page" do
-      get upload_path
+      get admin_upload_path
       expect(response).to be_successful
       expect(response.body).to include("Upload CSV")
     end
 
     it "uploads a csv file" do
-      post upload_path, params: { upload: { file: fixture_file_upload("test.csv") } }
+      post admin_upload_path, params: { upload: { file: fixture_file_upload("test.csv") } }
       expect(response).to be_successful
       expect(response.body).to include("Upload CSV Complete")
     end
 
     it "handles missing file" do
-      post upload_path, params: { upload: { file: nil } }
+      post admin_upload_path, params: { upload: { file: nil } }
       expect(response.body).to include("There is a problem")
     end
   end
